Related to #1190: do not patch if same content and split big tests
